// backend/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        Campaign: {
            id: string;
            name: string;
            vertical: string;
            /** @description JSON object of allowed regions/states */
            geo_rules?: unknown;
            supplier_id?: string | null;
            /** @enum {string} */
            status: "ACTIVE" | "INACTIVE";
            recording_default_enabled: boolean;
            /** Format: date-time */
            created_at: string;
            /** Format: date-time */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}